// src/app/modules/auth/models/register-response.model.ts
export class RegisterResponseModel {
  userId: number;
  userName: string;
  password: string;
  email: string;
  phoneNumber: string;
  userType: number;
  userStatus: number;
  gender: number;
  profileImage: string;
  year: string;
  nationalId: string;
  mobileToken: string;
  token: string;
  approveBy: string;
  approveById: string;
  approveAt: Date;
  creationDate: Date;
}